pyproject.toml
[tool.poetry]
name = "wriveted-api"
version = "0.4.1"
description = "Wriveted helps children find books they will love to read"
authors = ["Brian Thorne <brian@wriveted.com>", "Joshua Landy <joshua@wriveted.com>"]
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
alembic = "^1.11"
structlog = "^24.1.0"
psycopg2-binary = "^2.9.6"
SQLAlchemy = "^2.0"
httpx = "^0.28"
fastapi-cloudauth = "^0.4.3"
cryptography = "^44.0"
pydantic = {extras = ["email"], version = "^2.7"}
fastapi_permissions = "^0.2.7"
gunicorn = "^23.0"
tenacity = "^9.0"
sendgrid = "^6.10"
mock = "^5.0.1"
webencodings = "^0.5.1"
google-cloud-tasks = "^2.10.4"
# Google cloud tasks assumes setuptools but doesn't explicitly depend on it
setuptools = "^75.6"
humanize = "^4.4.0"
google-cloud-storage = "^2.6.0"
pillow = "^10.0"
stripe = "^11.4"
slack-sdk = "^3.19.5"
alembic-utils = "^0.8.1"
phonenumbers = "^8.13.6"
twilio = "^8.2.1"
xmltodict = "^0.13.0"
isbnlib = "^3.10.13"
openai = "^1.58"
google-genai = "^1.0"
google-api-python-client = "^2.156"
beautifulsoup4 = "^4.12.2"
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.30"
opentelemetry-api = "^1.29"
opentelemetry-sdk = "^1.29"
opentelemetry-exporter-gcp-trace = "^1.6.0"
opentelemetry-instrumentation-httpx = "^0.50b0"
opentelemetry-instrumentation-fastapi = "^0.50b0"
opentelemetry-instrumentation-sqlalchemy = "^0.50b0"
opentelemetry-instrumentation-vertexai = "^0.34"
opentelemetry-propagator-gcp = "^1.6.0"
opentelemetry-instrumentation-asyncpg = "^0.50b0"
opentelemetry-instrumentation-psycopg2 = "^0.50b0"
common-expression-language = "^0.5.0"
bleach = "^6.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8"
pytest-cov = "^5.0.0"
google-api-python-client = "^2.156"
google-auth-httplib2 = "^0.2"
google-auth-oauthlib = "^1.0"
pre-commit = "^4.0"
locust = "^2.32"
rich = "^13.3.2"
more-itertools = "^10.2.0"
pytest-asyncio = "^0.23.7"
pytest-timeout = "^2.3.1"
ruff = "^0.4.10"
mypy = "^1.16.0"
types-setuptools = "^80.9.0.20250529"
psutil = "^6.0"




[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
timeout = 300
cache_dir = "/tmp/pytest_cache"
addopts = [
    "--strict-markers",
    "--tb=short"
]
filterwarnings = [
    "ignore:.*FieldValidationInfo.*:DeprecationWarning",
    "ignore:.*CRUDCollection is deprecated.*:DeprecationWarning",
    "ignore:.*CRUDEvent is deprecated.*:DeprecationWarning",
    "ignore:.*pydantic.error_wrappers.*:UserWarning",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore:.*Pydantic serializer warnings.*:UserWarning",
    "ignore:.*Duplicate Operation ID.*:UserWarning",
    "ignore:.*You should not use the 'timeout' argument with the TestClient.*:DeprecationWarning",
    "ignore:.*conflicts with persistent instance.*:sqlalchemy.exc.SAWarning",
    "ignore:.*incompatible polymorphic identity.*:sqlalchemy.exc.SAWarning",
    "ignore:.*expected to delete .* rows.*:sqlalchemy.exc.SAWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests"
]

[tool.ruff]

line-length = 88
target-version = "py311"
exclude = [
    "examples",
    "scripts",
    "tests",
]

[tool.ruff.lint]

# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F"]
ignore = ["E501"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"app/**/__init__.py" = ["F401"]
"alembic/env.py" = ["E402"]
